Moved squared_distance(Point_3,Triangle_3
